Move RedisStorageManager to api_test.go
This was replaced by the cluster wrapper in the production code, but the
tests weren't updated. It should be removed once the tests have been
fixed. For now just move it to the test code, to make it clearer that
it's not actually used.
